Now can render heading and stripe seperate
